+++ src/game/Scene.h
@@ -29,8 +29,8 @@ namespace Engine
         void render(float frameExtrapolationTimeStep);
 
     private:
-        void loadMap(std::string_view tilesetFilename, std::string_view tilemapFilename, int tileWidth,
-                     int tileHeight, int tilesetColumns, float scale);
+        void loadMap(std::string_view tilesetFilename, std::string_view tilemapFilename, float tileWidth,
+                     float tileHeight, int tilesetColumns, float scale);
         void loadEntities();
         void loadLevel(int level);
         std::unique_ptr<Registry> m_registry{};
